scripts/linux: do not mv to-be-instrumented.lst when the second pass

+++ scripts/linux/__create_symlinks.sh
@@ -41,7 +41,7 @@ create_to_be_instrumented_functions_symlink() {
 	FILENAME="$TMP_DIR/to-be-instrumented-functions.lst"
 	SRC="$(__append_suffix $FILENAME)"
 
-	if [ -f "$TMP_DIR/kssb_rebuild" ]; then
+	if [ -f "$TMP_DIR/kssb_rebuild" -a -n "$_FIRSTPASS" ]; then
 		mv "$SRC" "$SRC".old || true
 	fi
